+++ lib/src/models/comment_impl.dart
@@ -6,16 +6,27 @@
 import 'dart:async';
 
 // Required for `hash2`
-import 'package:quiver/core.dart';
+import 'package:quiver/core.dart' show hash2;
 
 import 'package:draw/src/api_paths.dart';
 import 'package:draw/src/base_impl.dart';
 import 'package:draw/src/exceptions.dart';
+import 'package:draw/src/getter_utils.dart';
 import 'package:draw/src/reddit.dart';
 import 'package:draw/src/models/comment_forest.dart';
+import 'package:draw/src/models/redditor.dart';
 import 'package:draw/src/models/submission_impl.dart';
+import 'package:draw/src/models/subreddit.dart';
 import 'package:draw/src/models/user_content.dart';
 import 'package:draw/src/models/mixins/inboxable.dart';
+import 'package:draw/src/models/mixins/user_content_mixin.dart';
+import 'package:draw/src/models/mixins/editable.dart';
+import 'package:draw/src/models/mixins/gildable.dart';
+import 'package:draw/src/models/mixins/inboxtoggleable.dart';
+import 'package:draw/src/models/mixins/replyable.dart';
+import 'package:draw/src/models/mixins/reportable.dart';
+import 'package:draw/src/models/mixins/saveable.dart';
+import 'package:draw/src/models/mixins/voteable.dart';
 
 void setSubmissionInternal(commentLike, SubmissionRef s) {
   commentLike._submission = s;
@@ -39,7 +50,7 @@ void setRepliesInternal(commentLike, CommentForest comments) {
 
 /// Represents comments which have been collapsed under a 'load more comments'
 /// or 'continue this thread' section.
-class MoreComments extends RedditBase {
+class MoreComments extends RedditBase with RedditBaseInitializedMixin {
   static final RegExp _submissionRegExp = new RegExp(r'{id}');
   List<Comment> _comments;
   List<String> _children;
@@ -60,7 +71,9 @@ class MoreComments extends RedditBase {
       : _children = data['children'],
         _count = data['count'],
         _parentId = data['parent_id'],
-        super.loadData(reddit, data);
+        super(reddit) {
+    setData(this, data);
+  }
 
   bool operator ==(other) {
     return ((other is MoreComments) &&
@@ -99,13 +112,17 @@ class MoreComments extends RedditBase {
   }
 
   Future<Comment> _loadComment(String commentId) async {
+    if (_submission is! Submission) {
+      _submission = await _submission.populate();
+    }
+    final Submission initializedSubmission = _submission;
     final path = apiPath['submission'].replaceAll(
-            _submissionRegExp, fullnameSync(_submission).split('_')[1]) +
+            _submissionRegExp, initializedSubmission.fullname.split('_')[1]) +
         '_/' +
         commentId;
     final response = await reddit.get(path, params: {
-      'limit': (await _submission.property('commentLimit')),
-      'sort': (await _submission.property('commentSort'))
+      'limit': initializedSubmission.data['commentLimit'],
+      'sort': initializedSubmission.data['commentSort'],
     });
 
     final comments = response[1]['listing'];
@@ -116,6 +133,10 @@ class MoreComments extends RedditBase {
   /// Expand [MoreComments] into the list of actual [Comments] it represents.
   Future<List<Comment>> comments({bool update: true}) async {
     if (_comments == null) {
+      if (_submission is! Submission) {
+        _submission = await _submission.populate();
+      }
+      final Submission initializedSubmission = _submission;
       if (_count == 0) {
         return await _continueComments(update);
       }
@@ -123,7 +144,7 @@ class MoreComments extends RedditBase {
       assert(_children != null);
       final data = {
         'children': _children.join(','),
-        'link_id': fullnameSync(_submission),
+        'link_id': initializedSubmission.fullname,
         'sort': 'best', //(await _submission.property('commentSort')),
         'api_type': 'json',
       };
@@ -139,8 +160,145 @@ class MoreComments extends RedditBase {
   }
 }
 
-class Comment extends CommentRef {
+/// A fully initialized class which represents a single Reddit comment.
+class Comment extends CommentRef
+    with
+        EditableMixin,
+        GildableMixin,
+        InboxToggleableMixin,
+        InboxableMixin,
+        RedditBaseInitializedMixin,
+        ReplyableMixin,
+        ReportableMixin,
+        SaveableMixin,
+        UserContentMixin,
+        VoteableMixin {
   // CommentModeration get mod; // TODO(bkonyi): implement
+
+  /// Has this [Comment] been approved.
+  bool get approved => data['approved'];
+
+  /// When this [Comment] was approved.
+  ///
+  /// Returns `null` if this [Comment] has not been approved.
+  DateTime get approvedAtUtc =>
+      GetterUtils.dateTimeOrNull(data['approved_at_utc']);
+
+  /// Which Redditor approved this [Comment].
+  ///
+  /// Returns `null` if this [Comment] has not been approved.
+  RedditorRef get approvedBy =>
+      GetterUtils.redditorRefOrNull(reddit, data['approved_by']);
+
+  /// Is this [Comment] archived.
+  bool get archived => data['archived'];
+
+  // TODO(bkonyi): update this definition.
+  // RedditorRef get author => reddit.redditor(data['author']);
+
+  /// The author's flair text, if set.
+  ///
+  /// Returns `null` if the author does not have any flair text set.
+  String get authorFlairText => data['author_flair_text'];
+
+  /// When this [Comment] was removed.
+  ///
+  /// Returns `null` if the [Comment] has not been removed.
+  DateTime get bannedAtUtc => GetterUtils.dateTimeOrNull(data['banned_at_utc']);
+
+  /// Which Redditor removed this [Comment].
+  ///
+  /// Returns `null` if the [Comment] has not been removed.
+  RedditorRef get bannedBy =>
+      GetterUtils.redditorRefOrNull(reddit, data['banned_by']);
+
+  /// Is this [Comment] eligible for Reddit Gold.
+  bool get canGild => data['can_gild'];
+
+  bool get canModPost => data['can_mod_post'];
+
+  /// Is this [Comment] and its children collapsed.
+  bool get collapsed => data['collapsed'];
+
+  /// The reason for this [Comment] being collapsed.
+  ///
+  /// Returns `null` if the [Comment] isn't collapsed or there is no reason set.
+  String get collapsedReason => data['collapsed_reason'];
+
+  /// The time this [Comment] was created.
+  DateTime get createdUtc => GetterUtils.dateTimeOrNull(data['created_utc']);
+
+  /// The depth of this [Comment] in the tree of comments.
+  int get depth => data['depth'];
+
+  /// The number of downvotes this [Comment] has received.
+  int get downvotes => data['downs'];
+
+  /// Has this [Comment] been edited.
+  bool get edited => data['edited'];
+
+  /// Has this [Comment] be given Reddit Gold.
+  int get gilded => data['gilded'];
+
+  /// Ignore reports for this [Comment].
+  ///
+  /// This is only visible to moderators on the [Subreddit] this [Comment] was
+  /// posted on.
+  bool get ignoreReports => data['ignore_reports'];
+
+  /// Did the currently authenticated [User] post this [Comment].
+  bool get isSubmitter => data['is_submitter'];
+
+  /// Does the currently authenticated [User] like this [Comment].
+  bool get likes => data['likes'];
+
+  /// The id of the [Submission] link.
+  ///
+  /// Takes the form of `t3_7czz1q`.
+  String get linkId => data['link_id'];
+
+  /// The number of reports made regarding this [Comment].
+  ///
+  /// This is only visible to moderators on the [Subreddit] this [Comment] was
+  /// posted on.
+  int get numReports => data['num_reports'];
+
+  /// The ID of the parent [Comment] or [Submission].
+  String get parentId => data['parent_id'];
+
+  String get permalink => data['permalink'];
+
+  String get removalReason => data['removal_reason'];
+
+  /// Has this [Comment] been removed.
+  bool get removed => data['removed'];
+
+  /// Has this [Comment] been saved.
+  bool get saved => data['saved'];
+
+  /// The score associated with this [Comment] (aka net-upvotes).
+  int get score => data['score'];
+
+  /// Is this score of this [Comment] hidden.
+  bool get scoreHidden => data['score_hidden'];
+
+  /// Is this [Comment] marked as spam.
+  bool get spam => data['spam'];
+
+  /// Has this [Comment] been stickied.
+  bool get stickied => data['stickied'];
+
+  /// The [Subreddit] this [Comment] was posted in.
+  SubredditRef get subreddit => reddit.subreddit(data['subreddit']);
+
+  /// The id of the [Subreddit] this [Comment] was posted in.
+  String get subredditId => data['subreddit_id'];
+
+  /// The type of the [Subreddit] this [Comment] was posted in.
+  String get subredditType => data['subreddit_type'];
+
+  /// The number of upvotes this [Comment] has received.
+  int get upvotes => data['ups'];
 
   /// Returns true if the current [Comment] is a top-level comment. A [Comment]
   /// is a top-level comment if its parent is a [Submission].
@@ -149,15 +307,16 @@ class Comment extends CommentRef {
     return (parentIdType == reddit.config.submissionKind);
   }
 
-  /// The ID of the parent [Comment] or [Submission].
-  String get parentId => data['parent_id'];
-
   /// Return the parent of the comment.
   ///
   /// The returned parent will be an instance of either [Comment] or
   /// [Submission].
   Future<UserContent> parent() async {
-    if (parentId == await (await submission.populate()).fullname) {
+    if (_submission is! Submission) {
+      _submission = await _submission.populate();
+    }
+    final Submission initializedSubmission = _submission;
+    if (parentId == initializedSubmission.fullname) {
       return submission;
     }
 
@@ -186,7 +345,9 @@ class Comment extends CommentRef {
         ' lazy-comment');
   }
 
-  Comment.parse(Reddit reddit, Map data) : super.loadData(reddit, data);
+  Comment.parse(Reddit reddit, Map data) : super.withID(reddit, data['id']) {
+    setData(this, data);
+  }
 
   /// The [Submission] which this comment belongs to.
   SubmissionRef get submission {
@@ -197,15 +358,13 @@ class Comment extends CommentRef {
   }
 }
 
-/// A class which represents a single Reddit comment.
-class CommentRef extends UserContent with InboxableMixin {
+/// A lazily initialized class which represents a single Reddit comment. Can be
+/// promoted to a [Comment].
+class CommentRef extends UserContent {
   SubmissionRef _submission;
   CommentForest _replies;
 
   static final RegExp _commentRegExp = new RegExp(r'{id}');
-
-  CommentRef.loadData(Reddit reddit, Map data)
-      : super.loadDataWithPath(reddit, data, _infoPath(data['id']));
 
   CommentRef.withID(Reddit reddit, String id)
       : super.withPath(reddit, _infoPath(id));