+++ lib/src/models/inbox.dart
@@ -14,6 +14,7 @@ import 'package:draw/src/listing/listing_generator.dart';
 import 'package:draw/src/models/comment.dart';
 import 'package:draw/src/models/message.dart';
 
+/// A utility class used to interact with the Reddit inbox.
 class Inbox extends RedditBase {
   static final _messagesRegExp = new RegExp(r'{id}');
 
@@ -77,7 +78,7 @@ class Inbox extends RedditBase {
     messages.add(message);
     messages.addAll(message.replies);
     for (final m in messages) {
-      if (await m.fullname == messageId) {
+      if (m.fullname == messageId) {
         return m;
       }
     }