+++ lib/src/models/message.dart
@@ -7,16 +7,30 @@ import 'package:draw/src/base_impl.dart';
 import 'package:draw/src/models/mixins/inboxable.dart';
 import 'package:draw/src/reddit.dart';
 
-class Message extends RedditBase with InboxableMixin {
+/// A fully initialized class which represents a Message from the [Inbox].
+class Message extends RedditBase
+    with InboxableMixin, RedditBaseInitializedMixin {
   var _replies;
 
-  Message.parse(Reddit reddit, Map data) : super.loadData(reddit, data);
+  Message.parse(Reddit reddit, Map data) : super(reddit) {
+    setData(this, data);
+  }
 
   /// The author of the [Message].
   String get author => data['author'];
 
   /// The body of the [Message].
   String get body => data['body'];
+
+  /// When was this [Message] created.
+  DateTime get createdUtc => new DateTime.fromMillisecondsSinceEpoch(
+      data['created_utc'].round() * 1000,
+      isUtc: true);
+
+  /// Who is this [Message] for.
+  ///
+  /// Can be for either a [Redditor] or [Subreddit].
+  String get destination => data['dest'];
 
   /// The [List] of replies to this [Message].
   ///