+++ lib/src/models/mixins/inboxable.dart
@@ -10,8 +10,8 @@ import 'package:draw/src/base_impl.dart';
 import 'package:draw/src/reddit.dart';
 
 /// A mixin containing inbox functionality.
-abstract class InboxableMixin implements RedditBase {
-  Future<String> get fullname;
+abstract class InboxableMixin implements RedditBaseInitializedMixin {
+  String get fullname;
   Reddit get reddit;
 
   /// Returns true if the [Inbox] item is new.
@@ -29,8 +29,7 @@ abstract class InboxableMixin implements RedditBase {
   ///
   /// Note: Reddit does not permit blocking users unless you have a [Comment] or
   /// [Message] from them in your inbox.
-  Future block() async =>
-      await reddit.post(apiPath['block'], {'id': await fullname});
+  Future block() async => await reddit.post(apiPath['block'], {'id': fullname});
 
   /// Mark the item as collapsed.
   ///