+++ lib/src/models/mixins/saveable.dart
@@ -5,23 +5,23 @@
 
 import 'dart:async';
 
-import '../../api_paths.dart';
-import '../../reddit.dart';
+import 'package:draw/src/api_paths.dart';
+import 'package:draw/src/reddit.dart';
 
 /// Mixin for ReddieBase classes that can be saved.
 abstract class SaveableMixin {
   Reddit get reddit;
-  Future<String> get fullname;
+  String get fullname;
 
   /// Save the object.
   ///
   /// [category] (Gold only) is the category to save the object to. If your user does not
   /// have gold, this value is ignored.
-  Future save({String category}) async => reddit.post(
-      apiPath['save'], {'category': category ?? '', 'id': await fullname},
-      discardResponse: true);
+  Future save({String category}) async =>
+      reddit.post(apiPath['save'], {'category': category ?? '', 'id': fullname},
+          discardResponse: true);
 
   /// Unsave the object.
-  Future unsave() async => reddit
-      .post(apiPath['unsave'], {'id': await fullname}, discardResponse: true);
+  Future unsave() async =>
+      reddit.post(apiPath['unsave'], {'id': fullname}, discardResponse: true);
 }