+++ test/submission/submission_test.dart
@@ -93,8 +93,9 @@ Future main() async {
   test('lib/submission/reply', () async {
     final reddit = await createRedditTestInstance(
         'test/submission/lib_submission_reply.json');
-    final submission = new SubmissionRef.withPath(reddit,
-        r'https://www.reddit.com/r/drawapitesting/comments/7x6ew7/draw_using_dart_to_moderate_reddit_comments/');
+    final submission = await (new SubmissionRef.withPath(reddit,
+            r'https://www.reddit.com/r/drawapitesting/comments/7x6ew7/draw_using_dart_to_moderate_reddit_comments/'))
+        .populate();
     await submission.reply('Woohoo!');
   });
 }